requirements.txt
matplotlib==3.9.0
numpy==2.0.0
Pillow==10.3.0
